# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BMLT",
        "description": "BMLT Admin API Documentation",
        "license": {
            "name": "MIT",
            "url": "https://github.com/bmlt-enabled/bmlt-server/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/errortest": {
            "post": {
                "tags": [
                    "server"
                ],
                "summary": "Tests some errors",
                "description": "Tests some errors.",
                "operationId": "createErrorTest",
                "requestBody": {
                    "description": "Pass in error test object.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ErrorTest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Returns when POST is successful.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorTest"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Returns when user is not authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServerError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bmltToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorTest": {
                "properties": {
                    "arbitrary_string": {
                        "type": "string",
                        "example": "string"
                    },
                    "arbitrary_int": {
                        "type": "integer",
                        "example": "123"
                    },
                    "force_server_error": {
                        "type": "boolean",
                        "example": "true"
                    }
                },
                "type": "object"
            },
            "AuthenticationError": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "required": [
                    "message",
                    "errors"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The field is required. (and 1 more error)"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "error details"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "ServerError": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Server Error"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bmltToken": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "tokenUrl": "api/v1/auth/token",
                        "refreshUrl": "api/v1/auth/refresh",
                        "scopes": {}
                    }
                }
            }
        }
    },
...
}